feat: ajuste de adapter e conexao banco

+++ src/main/java/com/crecheconecta/escolar/application/model/Pagina.java
@@ -2,8 +2,7 @@ package com.crecheconecta.escolar.application.model;
 
 import java.util.List;
 
-public record Pagina<T>(List<T> itens, int pagina, int tamanho, int totalElementos, int totalPaginas) {
-
+public record Pagina<T>(List<T> itens, int pagina, int tamanho, long totalElementos, int totalPaginas) {
     public Pagina{
         itens = List.copyOf(itens);
     }
